Move tree removal to Filesystem; log and document the sweep
`Slot.remove_tree` was a slot method the supervisor also called, behind
an instance wrapper. The changelog entries ran to paragraphs, the README's
host launch swept the developer's `/tmp`, and `scratch.unswept` was not
in `docs/LOGS.md`.

Move the removal to `HotCell::Filesystem` and call it directly. Cut the
changelog entries to a sentence each, give the Procfile.dev cell a
`TMPDIR`, and document `scratch.unswept`.
