+++ README.md
@@ -299,10 +299,12 @@ Add an entry for your cell to `Procfile.dev`:
 
 ```procfile
 web: HOTCELL_ROOT=$PWD/tmp/hotcell-sockets bin/rails server
-cell: BUNDLE_GEMFILE=$PWD/hotcell/Gemfile HOTCELL_CONFIG=$PWD/hotcell/config.rb HOTCELL_OPERATIONS=$PWD/hotcell/operations HOTCELL_DIR=$PWD/tmp/hotcell-sockets/active_storage bundle exec hotcell
+cell: mkdir -p $PWD/tmp/hotcell-scratch && TMPDIR=$PWD/tmp/hotcell-scratch BUNDLE_GEMFILE=$PWD/hotcell/Gemfile HOTCELL_CONFIG=$PWD/hotcell/config.rb HOTCELL_OPERATIONS=$PWD/hotcell/operations HOTCELL_DIR=$PWD/tmp/hotcell-sockets/active_storage bundle exec hotcell
 ```
 
-Then `bin/dev` boots both, and the app finds the sockets under `tmp/hotcell-sockets`.
+Then `bin/dev` boots both, and the app finds the sockets under `tmp/hotcell-sockets`. `TMPDIR` matters:
+the cell empties its temporary directory of everything it owns at boot, and without one that is your
+`/tmp`. The directory has to exist, because Ruby falls back to `/tmp` when it does not.
 
 #### Configure the cell and operation limits
