+++ hotcell-server/lib/hot_cell/server.rb
@@ -8,6 +8,7 @@ require "hot_cell/limits"
 require "hot_cell/configuration"
 require "hot_cell/registry"
 require "hot_cell/operation"
+require "hot_cell/filesystem"
 require "hot_cell/slot"
 require "hot_cell/log"
 require "hot_cell/timing"
